Master - my very first
I really have no idea what I'm doing here. I'm trying to follow along with the prompts to practice creating a branch to my master. The prompts talked about a file dropdown in which I could branch out from the master but that didn't exist. So I started creating a new file and guess what - I'm being ask to create a master - so that's good! What a happy coincidence.